+++ setup.cfg
@@ -45,6 +45,7 @@ install_requires =
   uvicorn >= 0.30.5
   pytest-asyncio
   setuptools
+  tenacity
   playwright-stealth
 
 package_dir=
