+++ src/truflation/data/connector/cache.py
@@ -1,0 +1,33 @@
+from .base import Connector
+
+class ConnectorCache(Connector):
+    def __init__(self, cache, default_key=None):
+        super().__init__()
+        self.default_key = default_key
+        self.cache = cache
+
+    def read_all(self, *args, **kwargs):
+        key = kwargs.get('key', self.default_key)
+        return self.cache.get(key) if key is not None else None
+
+    def write_all(self, value, *args, **kwargs):
+        key = kwargs.get('key', self.default_key)
+        self.cache.set(key, value)
+
+
+
+class Cache:
+    def __init__(self):
+        self.cache_data = {}
+
+    def set(self, key, value):
+        self.cache_data[key] = value
+
+    def get(self, key):
+        return self.cache_data[key]
+
+    def connector(self, default_key=None):
+        return ConnectorCache(self, default_key)
+
+    def clear(self):
+        self.cache_data.clear()
